+++ KeppyDriverConfigurator/KeppyDriverConfigurator/KeppyDriverSFListAssign.cs
@@ -17,6 +17,7 @@ namespace KeppyDriverConfigurator
 
         public static string soundfontnewlocation = System.Environment.GetEnvironmentVariable("USERPROFILE").ToString();
 
+        public string ListsPath = soundfontnewlocation + "\\Keppy's Driver\\applists";
         public string List1Path = soundfontnewlocation + "\\Keppy's Driver\\applists\\keppymidi.applist";
         public string List2Path = soundfontnewlocation + "\\Keppy's Driver\\applists\\keppymidib.applist";
         public string List3Path = soundfontnewlocation + "\\Keppy's Driver\\applists\\keppymidic.applist";
@@ -78,30 +79,35 @@ namespace KeppyDriverConfigurator
         {
             try
             {
+                if (!System.IO.Directory.Exists(ListsPath))
+                {
+                    Directory.CreateDirectory(ListsPath);
+                    File.Create(WhichList).Dispose();
+                    Lis.Items.Clear();
+                    return;
+                }
                 if (!System.IO.File.Exists(WhichList))
                 {
                     File.Create(WhichList).Dispose();
                     Lis.Items.Clear();
+                    return;
                 }
-                else
+                try
                 {
-                    try
+                    using (StreamReader r = new StreamReader(WhichList))
                     {
-                        using (StreamReader r = new StreamReader(WhichList))
+                        string line;
+                        Lis.Items.Clear();
+                        while ((line = r.ReadLine()) != null)
                         {
-                            string line;
-                            Lis.Items.Clear();
-                            while ((line = r.ReadLine()) != null)
-                            {
-                                Lis.Items.Add(line); // The program is copying the entire text file to the List I's listbox.
-                            }
+                            Lis.Items.Add(line); // The program is copying the entire text file to the List I's listbox.
                         }
                     }
-                    catch
-                    {
-                        File.Create(WhichList).Dispose();
-                        MessageBox.Show("The soundfont list was missing, so the configurator automatically created it for you.", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                    }
+                }
+                catch
+                {
+                    File.Create(WhichList).Dispose();
+                    MessageBox.Show("The soundfont list was missing, so the configurator automatically created it for you.", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
                 }
             }
             catch (Exception ex)
